hotfix B2: payday window + slowdown + phase weights + £0.50 + weekday + even-spread traffic clear
- pacing.ts: 18-day payday window (Thu before last Fri → Mon+17); new
  onsale_slowdown / onsale_payday phases replace single onsale; PHASE_WEIGHTS
  export (presale=1.0, slowdown=0.6, payday=1.2, final10=1.5, event=2.0);
  ratios updated (final10/event → 25/75); phase-weighted per-day budget;
  roundHalf (£0.50) applied to all values, last day Conversion absorbs drift
- format.ts: fmtDayWithWeekday uses long weekday ("Saturday" not "Sat")
- event-plan-tab.tsx: buildEvenSpreadPatches deletes traffic key on every day;
  success banner adds "Traffic cleared."
- plan-header.tsx: Even spread confirm dialog shows £X/day + clear-Traffic notice

Made-with: Cursor

+++ lib/dashboard/format.ts
@@ -94,11 +94,11 @@ export function fmtDay(d: Date): string {
 }
 
 /**
- * Long date with short weekday — "19 Apr 2026 · Sat".
+ * Long date with full weekday name — "19 Apr 2026 · Saturday".
  *
  * Built in two locale calls (rather than one with both `weekday` and
- * `day` set) because Intl in en-GB renders weekday-first ("Sat 19 Apr
- * 2026") when both are requested; we want the date to lead so the
+ * `day` set) because Intl in en-GB renders weekday-first ("Saturday 19
+ * Apr 2026") when both are requested; we want the date to lead so the
  * column sorts visually like a date column.
  */
 export function fmtDayWithWeekday(d: Date): string {
@@ -107,7 +107,7 @@ export function fmtDayWithWeekday(d: Date): string {
     month: "short",
     year: "numeric",
   });
-  const weekday = d.toLocaleDateString("en-GB", { weekday: "short" });
+  const weekday = d.toLocaleDateString("en-GB", { weekday: "long" });
   return `${date} · ${weekday}`;
 }
